[ECO-5196] updated EventTypes.kt, Messages.kt, refactored integration tests - Added map for messageActionToEventType, used the same for emitting related message create, update and delete events - Refactored SandboxTest.kt, moved all of tests under `integration` package.
