+++ chat-android/src/test/java/com/ably/chat/integration/PresenceIntegrationTest.kt
@@ -1,0 +1,41 @@
+package com.ably.chat.integration
+
+import com.ably.chat.RoomOptions
+import java.util.UUID
+import kotlinx.coroutines.test.runTest
+import org.junit.Assert.assertEquals
+import org.junit.BeforeClass
+import org.junit.Test
+
+class PresenceIntegrationTest {
+
+    @Test
+    fun `should return empty list of presence members if nobody is entered`() = runTest {
+        val chatClient = sandbox.createSandboxChatClient()
+        val room = chatClient.rooms.get(UUID.randomUUID().toString(), RoomOptions.default)
+        room.attach()
+        val members = room.presence.get()
+        assertEquals(0, members.size)
+    }
+
+    @Test
+    fun `should return yourself as presence member after you entered`() = runTest {
+        val chatClient = sandbox.createSandboxChatClient("sandbox-client")
+        val room = chatClient.rooms.get(UUID.randomUUID().toString(), RoomOptions.default)
+        room.attach()
+        room.presence.enter()
+        val members = room.presence.get()
+        assertEquals(1, members.size)
+        assertEquals("sandbox-client", members.first().clientId)
+    }
+
+    companion object {
+        private lateinit var sandbox: Sandbox
+
+        @JvmStatic
+        @BeforeClass
+        fun setUp() = runTest {
+            sandbox = Sandbox.createInstance()
+        }
+    }
+}
