+++ chat-android/src/test/java/com/ably/chat/integration/TypingIntegrationTest.kt
@@ -1,0 +1,45 @@
+package com.ably.chat.integration
+
+import com.ably.chat.RoomOptions
+import com.ably.chat.TypingEvent
+import com.ably.chat.TypingOptions
+import java.util.UUID
+import kotlinx.coroutines.CompletableDeferred
+import kotlinx.coroutines.test.runTest
+import org.junit.Assert.assertEquals
+import org.junit.BeforeClass
+import org.junit.Test
+
+class TypingIntegrationTest {
+
+    @Test
+    fun `should return typing indication for client`() = runTest {
+        val chatClient1 = sandbox.createSandboxChatClient("client1")
+        val chatClient2 = sandbox.createSandboxChatClient("client2")
+        val roomId = UUID.randomUUID().toString()
+        val roomOptions = RoomOptions(typing = TypingOptions(timeoutMs = 10_000))
+        val chatClient1Room = chatClient1.rooms.get(roomId, roomOptions)
+        chatClient1Room.attach()
+        val chatClient2Room = chatClient2.rooms.get(roomId, roomOptions)
+        chatClient2Room.attach()
+
+        val deferredValue = CompletableDeferred<TypingEvent>()
+        chatClient2Room.typing.subscribe {
+            deferredValue.complete(it)
+        }
+        chatClient1Room.typing.start()
+        val typingEvent = deferredValue.await()
+        assertEquals(setOf("client1"), typingEvent.currentlyTyping)
+        assertEquals(setOf("client1"), chatClient2Room.typing.get())
+    }
+
+    companion object {
+        private lateinit var sandbox: Sandbox
+
+        @JvmStatic
+        @BeforeClass
+        fun setUp() = runTest {
+            sandbox = Sandbox.createInstance()
+        }
+    }
+}
